Properly mark mmap_ent as global
It currently happens to work fine using cg_llvm as the global_asm!()
ends up in the same cgu as the use of this global. There is no guarantee
that this will always be the case and in fact in case of cg_clif this
happens to not be the case, thus causing a linker error.