introduce proper formatting
This commit adds the `.editorconfig` and `.rustfmt.toml` file to the
root of the repository. With these file, we get

1. good IDE formatting support
2. a unified format for the whole crate

---

The formatting options were chosen to fit the current style of the code
present in this repository. They can be changed easily anytime in the
future.

---

Either with a follow-up commit, or with another PR, we can introduce a
GitHub actions workflow that lints the code for proper formatting.