Start integrating BIOS bootloader into build system
We can use an artifact dependency on the outer level since the x86_64-unknown-none target is Tier 2 and thus requires no -Zbuild-std.

For the boot_sector, we need to compile with a specific linker script and profile. We might need to duplicate the profile settings in the boot_sector Cargo.toml when publishing.