Limit BIOS bootloader's `max_phys_addr` to 4 GiB
When determining the maximum physical address for the BIOS bootloader's
identity mapping, we currently use the highest address in the E380
memory map, no matter how high it is. However, the bootloader runs in
protected mode, and therefore, it cannot address more than 4 GiB of
memory. We can save some time by not identity mapping addresses over 4
GiB, as the bootloader cannot address them anyway.

This commit changes the BIOS bootloader to skip addresses over 4 GiB
when determining the maximum physical address.

This is one of the changes described in issue #259.