Remove dependency on `time` crate
Version 0.1 of the `time` crate has a vulnerability. Apparently it does not affect the `chrono` crate, which is the crate that depends on `time`, but it's better to be safe. We don't actually need the default `chrono` feature of `fatfs`, so we can easily remove both `chrono` and `time` from our dependency tree by disabling that feature.