Print QEMU output directly instead of waiting until it finishes
The `output` method captures the output in a `Vec`, so we can only print it when the QEMU command is finished. This is problematic when a test panics inside the bootloader because we never exit QEMU in that case. Thus, we never print the error message, which makes debugging more difficult. This commit fixes this issue by piping the stdout and stderr directly to a background thread, which prints them directly.

We also print the full QEMU command and a separator line between test runs now, which should make the output easier to read.