Fix invalid mapping to zero page caused by off-by-one bug
The `zero_end` bound is exclusive, but we treat the `end_page` as inclusive. So when `zero_end` is page-aligned, we allocate one additional bss page. If this page was already mapped to some other segment, we remap it to a page with random content.

This is the same bug as https://github.com/rust-osdev/bootloader/pull/362.