avoid 32-bit relocation to __BOOTLOADER_CONFIG
As explained in my comment in #427, the compiler seems to have trouble
emitting relocations for references to global variables in custom
sections. For custom sections, it always emits 32-bit relocations even
when a 64-bit relocation would be required. This patch works around
that by never referencing the global in the custom section directly
from code, but only through a pointer from another global variable in
the non-custom .data section. The relocation used for the pointer in
the global variable will always use a 64-bit relocation.