don't call slice::as_ptr
We want to avoid directly accessing the data in the .bootloader-config
section and slice::as_ptr gets a pointer to that data. Instead we cast
a reference to a reference to the data in .bootloader-config to an
address and pass that to the asm! block. The problem with as_ptr was
that it's a method on the slice and not the reference and so by
calling slice::as_ptr, we once again directly accessed the slice when
we only meant to access the reference to the reference to the data.