construct memory in memory mapped by other page table
This has the advantage that we don't need to map the memory map into
the bootloader's page tables. Up until now, we've assumed that UEFI
only maps memory into regions covered by the first PML4 entry, but it
turns out that some implementations map the frame buffer into memory
mapped by other PML4 entries (this is totally legal, the frame buffer
need not be identity mapped). Previously we removed all but the first
PML4 entry, so that we can map our own memory there, but it's now clear
that this can lead to problems. The solution to this is to not modify
the page tables constructed by UEFI and keep all its PML4 entries, but
this is problematic because the code constructing the boot info and
memory map assumes that it can map them into the same addresses used
for the kernel's address space, but UEFI may have already mapped some
memory in there. Instead of mapping the boot info and memory map into
the bootloader's address space, we now only map them into the kernel's
address space. When we want to write to them, we first look up the
physical addresses in the kernel's page table and write to the identity
mapping. RemoteMemoryRegion implements this.
We have some unit tests for constructing the memory map and we want
those unit tests to be able to run outside the bootloader, so we can't
use RemoteMemoryRegion. To solve this, we introduce a trait
MemoryRegionSlice that only requires implementing a method for writing
a memory region and implement that slice for MemoryRegionSlice as well
as MemoryRegionSlice.