use threads instead of futures in build.rs
There's no inherent reason why we should use future-based concurrency
in build.rs. We don't spawn many futures and we don't make use of any
of the nice patterns made possible by structured concurrency. The
downside of using futures is that we have to pull in a bunch of
dependencies. Switching to sync code and removing the dependencies
resulted in a nice performance improvement (35s -> 30s). Also, it's just
good practice to reduce the number of unneeded dependencies especially
if some of them are unmaintained.