remove symbol mangling hack
Now that we fixed the buffer overflow and found another way of
generating smaller binaries, we no longer need to use the old symbol
mangling scheme.
The new symbol mangling was never really at fault. All it did was
shuffle some of the symbols around resulting in a larger binary. We
just got unlucky there.